pyproject.toml
[build-system]
requires = ["hatchling>=1.24", "uv-dynamic-versioning>=0.7.0"]
build-backend = "hatchling.build"

[project]
name = "et-mlapi"
dynamic = ["version"]
description = "Production-ready ML API template with Robyn — typed routing, event-driven architecture, WebSockets, SSE, streaming"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "Damien Volkov", email = "damvolkov@proton.me" }]
keywords = ["robyn", "ml", "api", "template", "machine-learning", "pydantic", "websockets", "sse"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Framework :: AsyncIO",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]

dependencies = [
    "robyn>=0.72.0",
    "pydantic>=2.11.0",
    "pydantic-settings[yaml]>=2.10.0",
    "orjson>=3.10.0",
    "structlog>=25.0.0",
    "beartype>=0.21.0",
    "gitpython>=3.1.45",
    "httpx>=0.28.0",
]

[project.scripts]
et-mlapi = "et_mlapi.main:main"

[project.urls]

[dependency-groups]
dev = [
    "pytest>=8.4.0",
    "pytest-asyncio>=1.0.0",
    "pytest-cov>=7.0.0",
    "pytest-mock>=3.14.0",
    "pytest-xdist>=3.5.0",
    "pytest-rerunfailures>=15.0",
    "ruff>=0.14.0",
    "ty>=0.0.5",
    "pre-commit>=4.4.0",
    "respx>=0.22.0",
    "polyfactory>=3.3.0",
]

[tool.hatch.version]
source = "uv-dynamic-versioning"

[tool.hatch.build.targets.wheel]
packages = ["src/et_mlapi"]

[tool.uv-dynamic-versioning]
vcs = "git"
style = "pep440"
bump = true

[tool.hatch.envs.default]
installer = "uv"

[tool.hatch.envs.default.scripts]
dev = "python -m et_mlapi.main"
lint = "ruff check --fix ."
format = "ruff format ."
typecheck = "ty check"
test = "python -m pytest tests/unit -v -m 'not slow' --cov --cov-report=term-missing"
test-integration = "python -m pytest tests/integration -v -m slow"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
testpaths = ["tests"]
filterwarnings = ["ignore::DeprecationWarning"]
markers = ["slow: marks tests as slow (deselect with '-m \"not slow\"')"]
addopts = "--import-mode=importlib"

[tool.ruff]
line-length = 120
target-version = "py312"
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "C4", "SIM"]
ignore = ["E501"]

[tool.ruff.lint.isort]
known-first-party = ["et_mlapi"]

[tool.ty.environment]
python-version = "3.12"
root = ["./src/et_mlapi"]
extra-paths = ["."]

[tool.ty.rules]
possibly-unresolved-reference = "error"
unresolved-import = "error"

[tool.coverage.run]
source = ["et_mlapi"]
branch = true
omit = [
    "src/et_mlapi/main.py",
]

[tool.coverage.report]
fail_under = 90
show_missing = true
exclude_lines = [
    "pragma: no cover",
    "if TYPE_CHECKING:",
    "raise NotImplementedError",
    "@abstractmethod",
    "@overload",
]
